Added delete to database. Fix error post was put.
